+++ include/gqmps2/detail/mpogen/mpogen_impl.h
@@ -89,7 +89,7 @@ template <typename TenElemType>
 void MPOGenerator<TenElemType>::AddTerm(
     const TenElemType coef,
     const GQTensorVec &local_ops,
-    const std::vector<size_t> &local_ops_idxs
+    const std::vector<int> &local_ops_idxs
 ) {
   assert(local_ops.size() == local_ops_idxs.size());
   assert(std::is_sorted(local_ops_idxs.cbegin(), local_ops_idxs.cend()));
@@ -97,10 +97,10 @@ void MPOGenerator<TenElemType>::AddTerm(
   if (coef == TenElemType(0)) { return; }   // If coef is zero, do nothing.
 
   auto coef_label = coef_label_convertor_.Convert(coef);
-  size_t ntrvl_ops_idxs_head = local_ops_idxs.front();
-  size_t ntrvl_ops_idxs_tail = local_ops_idxs.back();
+  int ntrvl_ops_idxs_head = local_ops_idxs.front();
+  int ntrvl_ops_idxs_tail = local_ops_idxs.back();
   OpReprVec ntrvl_ops_reprs;
-  for (size_t i = ntrvl_ops_idxs_head; i <= ntrvl_ops_idxs_tail; ++i) {
+  for (int i = ntrvl_ops_idxs_head; i <= ntrvl_ops_idxs_tail; ++i) {
     auto poss_it = std::find(local_ops_idxs.cbegin(), local_ops_idxs.cend(), i);
     if (poss_it != local_ops_idxs.cend()) {     // Nontrivial operator
       auto local_op_loc = poss_it - local_ops_idxs.cbegin();    // Location of the local operator in the local operators list.
@@ -146,9 +146,9 @@ template <typename TenElemType>
 void MPOGenerator<TenElemType>::AddTerm(
     const TenElemType coef,
     const GQTensorVec &phys_ops,
-    const std::vector<size_t> &phys_ops_idxs,
+    const std::vector<int> &phys_ops_idxs,
     const GQTensorVec &inst_ops,
-    const std::vector<std::vector<size_t>> &inst_ops_idxs_set
+    const std::vector<std::vector<int>> &inst_ops_idxs_set
 ) {
   assert(phys_ops.size() >= 2);
   assert(phys_ops.size() == phys_ops_idxs.size());
@@ -161,12 +161,12 @@ void MPOGenerator<TenElemType>::AddTerm(
   }
 
   GQTensorVec local_ops;
-  std::vector<size_t> local_ops_idxs;
-  for (size_t i = 0; i < phys_ops.size()-1; ++i) {
+  std::vector<int> local_ops_idxs;
+  for (int i = 0; i < phys_ops.size()-1; ++i) {
     local_ops.push_back(phys_ops[i]);
     local_ops_idxs.push_back(phys_ops_idxs[i]);
     if (inst_ops_idxs_set == kNullIntVecVec) {
-      for (size_t j = phys_ops_idxs[i]+1; j < phys_ops_idxs[i+1]; ++j) {
+      for (int j = phys_ops_idxs[i]+1; j < phys_ops_idxs[i+1]; ++j) {
         local_ops.push_back(inst_ops[i]);
         local_ops_idxs.push_back(j);
       }
@@ -183,7 +183,7 @@ void MPOGenerator<TenElemType>::AddTerm(
   local_ops_idxs.push_back(phys_ops_idxs.back());
   if (inst_ops.size() == phys_ops.size()) {
     if (inst_ops_idxs_set == kNullIntVecVec) {
-      for (size_t j = phys_ops_idxs.back(); j < N_; ++j) {
+      for (int j = phys_ops_idxs.back(); j < N_; ++j) {
         local_ops.push_back(inst_ops.back());
         local_ops_idxs.push_back(j);
       }
@@ -216,15 +216,15 @@ template <typename TenElemType>
 void MPOGenerator<TenElemType>::AddTerm(
     const TenElemType coef,
     const GQTensorT &op1,
-    const size_t op1_idx,
+    const int op1_idx,
     const GQTensorT &op2,
-    const size_t op2_idx,
+    const int op2_idx,
     const GQTensorT &inst_op,
-    const std::vector<size_t> &inst_op_idxs
+    const std::vector<int> &inst_op_idxs
 ) {
   if (op2 == GQTensorT()) {     // One-body interaction term
     GQTensorVec local_ops = {op1};
-    std::vector<size_t> local_ops_idxs = {op1_idx};
+    std::vector<int> local_ops_idxs = {op1_idx};
     AddTerm(coef, local_ops, local_ops_idxs);     // Use the most generic API
   } else {                      // Two-body interaction term
     assert(op2_idx != 0);
@@ -243,183 +243,6 @@ void MPOGenerator<TenElemType>::AddTerm(
     }
   }
 }
-
-
-/** MPOGenerator<TenElemType>::AddTerm
- * insertion operators fill the among physical operators compactly, usually for uniform cases.
- * The number of physical operators = 2 or more, but must specify the insertion operators
- * @tparam TenElemType GQTEN_Double or GQTEN_Complex
- * @param coef coefficient of physical term
- * @param phys_ops physical operators
- * @param idxs site numbers of physical operators
- * @param inst_ops insertion operator set
- */
-//template <typename TenElemType>
-//void MPOGenerator<TenElemType>::AddTerm(
-    //const TenElemType coef,
-    //const GQTensorVec &phys_ops,
-    //const std::vector<long> &idxs,
-    //const GQTensorVec &inst_ops) {
-  //assert(phys_ops.size() == idxs.size());
-  //for (auto idx : idxs) { assert(idx < N_); }
-  //assert((inst_ops.size() == phys_ops.size()-1) ||
-         //(inst_ops.size() == phys_ops.size()));
-  //if (coef == TenElemType(0)) { return; }   // If coef is zero, do nothing.
-  //CoefLabel coef_label = coef_label_convertor_.Convert(coef);
-  //long ntrvl_op_idx_head, ntrvl_op_idx_tail;
-  //ntrvl_op_idx_head = idxs.front();
-  //if (inst_ops.size() == phys_ops.size()-1) {
-    //ntrvl_op_idx_tail = idxs.back();
-  //} else if (inst_ops.size() == phys_ops.size()) {
-    //ntrvl_op_idx_tail = N_ - 1;
-  //}
-  //OpReprVec ntrvl_op_reprs;
-  //size_t last_phys_op_idx;
-  //for (long i = ntrvl_op_idx_head; i <= ntrvl_op_idx_tail; ++i) {
-    //auto poss_it = std::find(idxs.cbegin(), idxs.cend(), i);
-    //if (poss_it != idxs.cend()) {
-      //auto phys_ops_idx = poss_it - idxs.cbegin();
-      //OpLabel op_label = op_label_convertor_.Convert(phys_ops[phys_ops_idx]);
-      //if (phys_ops_idx == 0) {
-        //ntrvl_op_reprs.push_back(OpRepr(coef_label, op_label));
-      //} else {
-        //ntrvl_op_reprs.push_back(OpRepr(op_label));
-      //}
-      //last_phys_op_idx = phys_ops_idx;
-    //} else {
-      //OpLabel op_label = op_label_convertor_.Convert(
-                             //inst_ops[last_phys_op_idx]);
-      //ntrvl_op_reprs.push_back(OpRepr(op_label));
-    //}
-  //}
-  //assert(ntrvl_op_reprs.size() == (ntrvl_op_idx_tail - ntrvl_op_idx_head + 1));
-
-  //fsm_.AddPath(ntrvl_op_idx_head, ntrvl_op_idx_tail, ntrvl_op_reprs);
-//}
-
-
-/** MPOGenerator<TenElemType>::AddTerm
- * insertion operator sites are specified, can be used for non-uniform cases.
- * The number of physical operators = 2 or more, but must specify the insertion operators and their positions
- * @tparam TenElemType GQTEN_Double or GQTEN_Complex
- * @param coef coefficient
- * @param phys_ops physical operators
- * @param idxs corresponding site number of physical operators
- * @param inst_ops insertion operators, # = (# of idxs) or (# of idxs -1)
- * @param inst_idxs the number of which sites was inserted, usually for
- *          the non-uniform hilbert spaces cases.
- */
-//template <typename TenElemType>
-//void MPOGenerator<TenElemType>::AddTerm(
-        //const TenElemType coef,
-        //GQTensorVec phys_ops,
-        //std::vector<long> idxs,
-        //const GQTensorVec &inst_ops,
-        //const std::vector<long> &inst_idxs) {
-  //assert(phys_ops.size() == idxs.size());
-  //for (auto idx : idxs) { assert(idx < N_); }
-  //assert((inst_ops.size() == phys_ops.size()-1) ||
-         //(inst_ops.size() == phys_ops.size()));
-  //if (coef == TenElemType(0)) { return; }   // If coef is zero, do nothing.
-
-  //if (!std::is_sorted(idxs.begin(),idxs.end())){
-    //std::map<long, GQTensorT> TmpMap;
-    //for (int i=0; i < idxs.size(); i++) {
-      //TmpMap[idxs[i]] = phys_ops[i];
-    //}
-    //typename std::map<long,GQTensorT>::iterator iter;
-    //int i = 0;
-    //for (iter = TmpMap.begin(); iter!=TmpMap.end();iter++){
-      //idxs[i] = iter->first;
-      //phys_ops[i] = iter->second;
-      //i=i+1;
-    //}
-    //TmpMap.clear();
-  //}
-
-  //CoefLabel coef_label = coef_label_convertor_.Convert(coef);
-
-  //long ntrvl_op_idx_head, ntrvl_op_idx_tail;
-  //ntrvl_op_idx_head = idxs.front();
-  //if (inst_ops.size() == phys_ops.size()-1) {
-    //ntrvl_op_idx_tail = idxs.back();
-  //} else if (inst_ops.size() == phys_ops.size()) {
-    //ntrvl_op_idx_tail = N_ - 1;
-  //}
-
-  //OpReprVec ntrvl_op_reprs;
-  //size_t last_phys_op_idx;
-  //for (long i = ntrvl_op_idx_head; i <= ntrvl_op_idx_tail; ++i) {
-    //auto poss_it = std::find(idxs.cbegin(), idxs.cend(), i);
-    //if (poss_it != idxs.cend()) {
-      //auto phys_ops_idx = poss_it - idxs.cbegin();
-      //OpLabel op_label = op_label_convertor_.Convert(phys_ops[phys_ops_idx]);
-      //if (phys_ops_idx == 0) {
-        //ntrvl_op_reprs.push_back(OpRepr(coef_label, op_label));
-      //} else {
-        //ntrvl_op_reprs.push_back(OpRepr(op_label));
-      //}
-      //last_phys_op_idx = phys_ops_idx;
-    //} else if ((std::find(inst_idxs.cbegin(), inst_idxs.cend(), i) ) != inst_idxs.cend() ) {
-      //OpLabel op_label = op_label_convertor_.Convert(
-              //inst_ops[last_phys_op_idx]);
-      //ntrvl_op_reprs.push_back(OpRepr(op_label));
-    //} else {
-      //OpLabel op_label = op_label_convertor_.Convert(id_op_vector_[i]);
-      //ntrvl_op_reprs.push_back(OpRepr(op_label));
-    //}
-  //}
-  //assert(ntrvl_op_reprs.size() == (ntrvl_op_idx_tail - ntrvl_op_idx_head + 1));
-
-  //fsm_.AddPath(ntrvl_op_idx_head, ntrvl_op_idx_tail, ntrvl_op_reprs);
-//}
-
-/** MPOGenerator<TenElemType>::AddTerm
- * used for two cases: 1. @param inst_op are specify, used for 2 or more sites interaction with inst_op filled
- *                          in all of the sites between the physical operators, usually for uniform lattices.
- *                     2. @param inst_op are neglected, used for 2 or more sites interaction without inst_op between
- *                          the physical operators, can be used for uniform/non-uniform lattices.
- * @tparam TenElemType GQTEN_Double or GQTEN_Complex
- * @param coef coefficient
- * @param phys_ops physical operators
- * @param idxs site number of physical operators
- * @param inst_op only one operator, the insertion operators between different intervals are the same
- */
-//template <typename TenElemType>
-//void MPOGenerator<TenElemType>::AddTerm(
-    //const TenElemType coef,
-    //const GQTensorVec &phys_ops,
-    //const std::vector<long> &idxs,
-    //const GQTensorT &inst_op) {
-  //auto phys_ops_num = phys_ops.size();
-  //assert(phys_ops_num > 0);
-  //GQTensorVec inst_ops(phys_ops_num-1, inst_op);
-  //if (inst_op != GQTensor<TenElemType>()) {
-    //AddTerm(coef, phys_ops, idxs, inst_ops);
-  //}else{
-    //std::vector<long> inst_idxs_null;
-    //AddTerm(coef, phys_ops, idxs, inst_ops,inst_idxs_null);
-  //}
-//}
-
-/** MPOGenerator<TenElemType>::AddTerm
- * For one site terms, e.g. external field term, Hubbard U term
- * @tparam TenElemType GQTEN_Double or GQTEN_Complex
- * @param coef coefficient
- * @param phys_op physical operator
- * @param idx site number
- */
-//template <typename TenElemType>
-//void MPOGenerator<TenElemType>::AddTerm(
-    //const TenElemType coef,
-    //const GQTensorT &phys_op,
-    //const long idx) {
-  //AddTerm(
-      //coef,
-      //GQTensorVec({phys_op}),
-      //std::vector<long>({idx}),
-      //GQTensorVec({}));
-//}
 
 
 template <typename TenElemType>
