+++ tests/test_two_site_algo.cc
@@ -38,7 +38,7 @@ void RunTestTwoSiteAlgorithmNoiseCase(
 
 
 // Helpers
-inline void KeepOrder(size_t &x, size_t &y) {
+inline void KeepOrder(int &x, int &y) {
   if (x > y) {
     auto temp = y;
     y = x;
@@ -47,28 +47,28 @@ inline void KeepOrder(size_t &x, size_t &y) {
 }
 
 
-inline size_t coors2idx(
-    const size_t x, const size_t y, const size_t Nx, const size_t Ny) {
+inline int coors2idx(
+    const int x, const int y, const int Nx, const int Ny) {
 	return x * Ny + y;
 }
 
 
 
-inline size_t coors2idxSquare(
-    const size_t x, const size_t y, const size_t Nx, const size_t Ny) {
+inline int coors2idxSquare(
+    const int x, const int y, const int Nx, const int Ny) {
   return x * Ny + y;
 }
 
 
-inline size_t coors2idxHoneycomb(
-    const size_t x, const size_t y, const size_t Nx, const size_t Ny) {
+inline int coors2idxHoneycomb(
+    const int x, const int y, const int Nx, const int Ny) {
   return Ny * (x%Nx) + y%Ny;
 }
 
 
 // Test spin systems
 struct TestTwoSiteAlgorithmSpinSystem : public testing::Test {
-  size_t N = 6;
+  int N = 6;
 
   QN qn0 = QN({QNNameVal("Sz", 0)});
   Index pb_out = Index({
@@ -103,7 +103,7 @@ struct TestTwoSiteAlgorithmSpinSystem : public testing::Test {
 
 TEST_F(TestTwoSiteAlgorithmSpinSystem, 1DIsing) {
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec_6, qn0);
-  for (size_t i = 0; i < N-1; ++i) {
+  for (int i = 0; i < N-1; ++i) {
     dmpo_gen.AddTerm(1, {dsz, dsz}, {i, i+1});
   }
   auto dmpo = dmpo_gen.Gen();
@@ -130,7 +130,7 @@ TEST_F(TestTwoSiteAlgorithmSpinSystem, 1DIsing) {
 
   // Complex Hamiltonian.
   auto zmpo_gen = MPOGenerator<GQTEN_Complex>(site_vec_6, qn0);
-  for (size_t i = 0; i < N-1; ++i) {
+  for (int i = 0; i < N-1; ++i) {
     zmpo_gen.AddTerm(1, {zsz, zsz}, {i, i+1});
   }
   auto zmpo = zmpo_gen.Gen();
@@ -147,7 +147,7 @@ TEST_F(TestTwoSiteAlgorithmSpinSystem, 1DIsing) {
 
 TEST_F(TestTwoSiteAlgorithmSpinSystem, 1DHeisenberg) {
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec_6, qn0);
-  for (size_t i = 0; i < N-1; ++i) {
+  for (int i = 0; i < N-1; ++i) {
     dmpo_gen.AddTerm(1,   {dsz, dsz}, {i, i+1});
     dmpo_gen.AddTerm(0.5, {dsp, dsm}, {i, i+1});
     dmpo_gen.AddTerm(0.5, {dsm, dsp}, {i, i+1});
@@ -182,7 +182,7 @@ TEST_F(TestTwoSiteAlgorithmSpinSystem, 1DHeisenberg) {
 
   // Complex Hamiltonian
   auto zmpo_gen = MPOGenerator<GQTEN_Complex>(site_vec_6, qn0);
-  for (size_t i = 0; i < N-1; ++i) {
+  for (int i = 0; i < N-1; ++i) {
     zmpo_gen.AddTerm(1,   {zsz, zsz}, {i, i+1});
     zmpo_gen.AddTerm(0.5, {zsp, zsm}, {i, i+1});
     zmpo_gen.AddTerm(0.5, {zsm, zsp}, {i, i+1});
@@ -204,7 +204,7 @@ TEST_F(TestTwoSiteAlgorithmSpinSystem, 1DHeisenberg) {
 
 TEST_F(TestTwoSiteAlgorithmSpinSystem, 2DHeisenberg) {
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec_6, qn0);
-  std::vector<std::pair<size_t, size_t>> nn_pairs = {
+  std::vector<std::pair<int, int>> nn_pairs = {
       std::make_pair(0, 1), 
       std::make_pair(0, 2), 
       std::make_pair(1, 3), 
@@ -269,13 +269,13 @@ TEST_F(TestTwoSiteAlgorithmSpinSystem, 2DHeisenberg) {
 
 
 TEST_F(TestTwoSiteAlgorithmSpinSystem, 2DKitaevSimpleCase) {
-  size_t Nx = 4;
-  size_t Ny = 2;
-  size_t N1 = Nx*Ny;
+  int Nx = 4;
+  int Ny = 2;
+  int N1 = Nx*Ny;
   SiteVec site_vec(N1, pb_out);
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec, qn0);
-  for (size_t x = 0; x < Nx; ++x) {
-    for (size_t y = 0; y < Ny; ++y) {
+  for (int x = 0; x < Nx; ++x) {
+    for (int y = 0; y < Ny; ++y) {
       if (x % 2 == 1) {
         auto s0 = coors2idxHoneycomb(x, y, Nx, Ny);
         auto s1 = coors2idxHoneycomb(x-1, y+1, Nx, Ny);
@@ -307,8 +307,8 @@ TEST_F(TestTwoSiteAlgorithmSpinSystem, 2DKitaevSimpleCase) {
 
   // Complex Hamiltonian
   auto zmpo_gen = MPOGenerator<GQTEN_Complex>(site_vec, qn0);
-  for (size_t x = 0; x < Nx; ++x) {
-    for (size_t y = 0; y < Ny; ++y) {
+  for (int x = 0; x < Nx; ++x) {
+    for (int y = 0; y < Ny; ++y) {
       if (x % 2 == 1) {
         auto s0 = coors2idxHoneycomb(x, y, Nx, Ny);
         auto s1 = coors2idxHoneycomb(x-1, y+1, Nx, Ny);
@@ -354,16 +354,16 @@ TEST(TestTwoSiteAlgorithmNoSymmetrySpinSystem, 2DKitaevComplexCase) {
   double K = 1.0;
   double Gm = 0.1;
   double h = 0.1;
-  size_t Nx = 3, Ny = 4;
-  size_t N = Nx*Ny;
+  int Nx = 3, Ny = 4;
+  int N = Nx*Ny;
   SiteVec site_vec(N, idx_out);
   auto mpo_gen = MPOGenerator<TenElemType>(site_vec, zero_div);
   // H =   J * \Sigma_{<ij>} S_i*S_j
   //       K * \Sigma_{<ij>,c-link} Sc_i*Sc_j
   //		 Gm * \Sigma_{<ij>,c-link} Sa_i*Sb_j + Sb_i*Sa_j
   //     - h * \Sigma_i{S^z}
-  for (size_t x = 0; x < Nx; ++x) {
-    for (size_t y = 0; y < Ny; ++y) {
+  for (int x = 0; x < Nx; ++x) {
+    for (int y = 0; y < Ny; ++y) {
       // use the configuration '/|\' to traverse the square lattice
       // single site operator
       auto site0_num = coors2idx(x, y, Nx, Ny);
@@ -417,7 +417,7 @@ TEST(TestTwoSiteAlgorithmNoSymmetrySpinSystem, 2DKitaevComplexCase) {
   std::vector<Tensor*> mps(N);
   std::vector<long> stat_labs(N);
   auto was_up = false;
-  for (size_t i = 0; i < N; ++i) {
+  for (int i = 0; i < N; ++i) {
     if (was_up) {
       stat_labs[i] = 1;
       was_up = false;
@@ -441,7 +441,7 @@ TEST(TestTwoSiteAlgorithmNoSymmetrySpinSystem, 2DKitaevComplexCase) {
 
 // Test fermion models.
 struct TestTwoSiteAlgorithmTjSystem2U1Symm : public testing::Test {
-  size_t N = 4;
+  int N = 4;
   double t = 3.0;
   double J = 1.0;
   QN qn0 = QN({QNNameVal("N", 0), QNNameVal("Sz", 0)});
@@ -502,7 +502,7 @@ struct TestTwoSiteAlgorithmTjSystem2U1Symm : public testing::Test {
 
 TEST_F(TestTwoSiteAlgorithmTjSystem2U1Symm, 1DCase) {
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec_4, qn0);
-  for (size_t i = 0; i < N-1; ++i) {
+  for (int i = 0; i < N-1; ++i) {
     dmpo_gen.AddTerm(-t, dcdagup, i, dcup, i+1, df);
     dmpo_gen.AddTerm(-t, dcdagdn, i, dcdn, i+1, df);
     dmpo_gen.AddTerm(-t, dcup, i, dcdagup, i+1, df);
@@ -530,7 +530,7 @@ TEST_F(TestTwoSiteAlgorithmTjSystem2U1Symm, 1DCase) {
 
   // Complex Hamiltonian
   auto zmpo_gen = MPOGenerator<GQTEN_Complex>(site_vec_4, qn0);
-  for (size_t i = 0; i < N-1; ++i) {
+  for (int i = 0; i < N-1; ++i) {
     zmpo_gen.AddTerm(-t, zcdagup, i, zcup, i+1, zf);
     zmpo_gen.AddTerm(-t, zcdagdn, i, zcdn, i+1, zf);
     zmpo_gen.AddTerm(-t, zcup, i, zcdagup, i+1, zf);
@@ -550,7 +550,7 @@ TEST_F(TestTwoSiteAlgorithmTjSystem2U1Symm, 1DCase) {
 
 TEST_F(TestTwoSiteAlgorithmTjSystem2U1Symm, 2DCase) {
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec_4, qn0);
-  std::vector<std::pair<size_t, size_t>> nn_pairs = {
+  std::vector<std::pair<int, int>> nn_pairs = {
       std::make_pair(0, 1), 
       std::make_pair(0, 2), 
       std::make_pair(2, 3), 
@@ -653,15 +653,15 @@ TEST_F(TestTwoSiteAlgorithmTjSystem1U1Symm, RashbaTermCase) {
   double J = 1.0;
   double lamb = 0.03;
   auto ilamb = GQTEN_Complex(0, lamb);
-  size_t Nx = 3;
-  size_t Ny = 2;
-  size_t Ntot = Nx * Ny;
+  int Nx = 3;
+  int Ny = 2;
+  int Ntot = Nx * Ny;
   char BCx = 'p';
   char BCy = 'o';
   SiteVec site_vec(Ntot, pb_out);
   auto mpo_gen = MPOGenerator<GQTEN_Complex>(site_vec, qn0);
-  for (size_t x = 0; x < Nx; ++x) {
-    for (size_t y = 0; y < Ny; ++y) {
+  for (int x = 0; x < Nx; ++x) {
+    for (int y = 0; y < Ny; ++y) {
 
       if (!((BCx == 'o') && (x == Nx-1))) {
         auto s0 = coors2idxSquare(x, y, Nx, Ny);
@@ -735,9 +735,9 @@ TEST_F(TestTwoSiteAlgorithmTjSystem1U1Symm, RashbaTermCase) {
 
 
 struct TestTwoSiteAlgorithmHubbardSystem : public testing::Test {
-  size_t Nx = 2;
-  size_t Ny = 2;
-  size_t N = Nx * Ny;
+  int Nx = 2;
+  int Ny = 2;
+  int N = Nx * Ny;
   double t0 = 1.0;
   double t1 = 0.5;
   double U = 2.0;
@@ -829,8 +829,8 @@ struct TestTwoSiteAlgorithmHubbardSystem : public testing::Test {
 
 TEST_F(TestTwoSiteAlgorithmHubbardSystem, 2Dcase) {
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec, qn0);
-  for (size_t i = 0; i < Nx; ++i) {
-    for (size_t j = 0; j < Ny; ++j) {
+  for (int i = 0; i < Nx; ++i) {
+    for (int j = 0; j < Ny; ++j) {
       auto s0 = coors2idxSquare(i, j, Nx, Ny);
       dmpo_gen.AddTerm(U, dnupdn, s0);
 
@@ -893,8 +893,8 @@ TEST_F(TestTwoSiteAlgorithmHubbardSystem, 2Dcase) {
 
   // Complex Hamiltonian
   auto zmpo_gen = MPOGenerator<GQTEN_Complex>(site_vec, qn0);
-  for (size_t i = 0; i < Nx; ++i) {
-    for (size_t j = 0; j < Ny; ++j) {
+  for (int i = 0; i < Nx; ++i) {
+    for (int j = 0; j < Ny; ++j) {
       auto s0 = coors2idxSquare(i, j, Nx, Ny);
       zmpo_gen.AddTerm(U, znupdn, s0);
 
@@ -950,8 +950,8 @@ TEST_F(TestTwoSiteAlgorithmHubbardSystem, 2Dcase) {
 // Test non-uniform local Hilbert spaces system.
 // Kondo insulator, ref 10.1103/PhysRevB.97.245119,
 struct TestKondoInsulatorSystem : public testing::Test {
-  size_t Nx = 4;
-  size_t N = 2 * Nx;
+  int Nx = 4;
+  int N = 2 * Nx;
   double t = 0.25;
   double Jk = 1.0;
   double Jz = 0.5;
@@ -999,7 +999,7 @@ struct TestKondoInsulatorSystem : public testing::Test {
     Sp({0,1}) = 1.0;
     Sm({1,0}) = 1.0;
 
-    for(size_t i = 0; i < N; ++i){
+    for(int i = 0; i < N; ++i){
       if(i%2==0) pb_set[i] = pb_outE;   // even site is extended electron
       if(i%2==1) pb_set[i] = pb_outL;   // odd site is localized electron
     }
@@ -1009,14 +1009,14 @@ struct TestKondoInsulatorSystem : public testing::Test {
 TEST_F(TestKondoInsulatorSystem, doublechain) {
   SiteVec site_vec(pb_set);
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec, qn0);
-  for (size_t i = 0; i < N-2; i=i+2){
+  for (int i = 0; i < N-2; i=i+2){
     dmpo_gen.AddTerm(-t, bupcF, i, bupa, i+2);
     dmpo_gen.AddTerm(-t, bdnc, i, Fbdna, i+2);
     dmpo_gen.AddTerm( t, bupaF, i, bupc, i+2);
     dmpo_gen.AddTerm( t, bdna, i, Fbdnc, i+2);
     dmpo_gen.AddTerm(Jz, Sz, i+1, Sz, i+3);
   }
-  for (size_t i = 0; i < N; i=i+2){
+  for (int i = 0; i < N; i=i+2){
     dmpo_gen.AddTerm(Jk, sz, i, Sz, i+1);
     dmpo_gen.AddTerm(Jk/2, sp, i, Sm, i+1);
     dmpo_gen.AddTerm(Jk/2, sm, i, Sp, i+1);
@@ -1045,13 +1045,13 @@ TEST_F(TestKondoInsulatorSystem, doublechain) {
 // Test noised tow-site DMRG algorithm.
 // Electron-phonon interaction Holstein chain, ref 10.1103/PhysRevB.57.6376
 struct TestHolsteinChain : public testing::Test {
-  size_t L = 4;           // The number of electron
-  size_t Np = 3;          // per electron has Np pseudosite
+  int L = 4;           // The number of electron
+  int Np = 3;          // per electron has Np pseudosite
   double t = 1;         // electron hopping
   double g = 1;         // electron-phonon interaction
   double U = 8;         // Hubbard U
   double omega = 5;     // phonon on-site potential
-  size_t N = (1+Np)*L;     // The length of the mps/mpo
+  int N = (1+Np)*L;     // The length of the mps/mpo
 
   QN qn0 = QN({ QNNameVal("Nf", 0), QNNameVal("Sz", 0) });
   //Fermion(electron)
@@ -1125,7 +1125,7 @@ struct TestHolsteinChain : public testing::Test {
     n_a({0,0}) = 1;
     idB({0,0}) = 1; idB({1,1}) = 1;
     P0 = idB+(-n_a);
-    for(size_t i =0;i < N; ++i){
+    for(int i =0;i < N; ++i){
       if(i%(Np+1)==0) pb_set[i] = pb_outF; // even site is fermion
       else pb_set[i] = pb_outB; // odd site is boson
     }
@@ -1135,15 +1135,15 @@ struct TestHolsteinChain : public testing::Test {
 TEST_F(TestHolsteinChain, holsteinchain) {
   SiteVec site_vec(pb_set);
   auto dmpo_gen = MPOGenerator<GQTEN_Double>(site_vec, qn0);
-  for (size_t i = 0; i < N-Np-1; i=i+Np+1) {
+  for (int i = 0; i < N-Np-1; i=i+Np+1) {
     dmpo_gen.AddTerm(-t, bupcF, i, bupa, i+Np+1);
     dmpo_gen.AddTerm(-t, bdnc, i, Fbdna, i+Np+1);
     dmpo_gen.AddTerm( t, bupaF, i, bupc, i+Np+1);
     dmpo_gen.AddTerm( t, bdna, i, Fbdnc, i+Np+1);
   }
-  for (size_t i = 0; i < N; i=i+Np+1) {
+  for (int i = 0; i < N; i=i+Np+1) {
     dmpo_gen.AddTerm(U, Uterm, i);
-    for(size_t j = 0; j < Np; ++j) {
+    for(int j = 0; j < Np; ++j) {
       dmpo_gen.AddTerm(omega, (double)(pow(2,j))*n_a, i+j+1);
     }
     dmpo_gen.AddTerm(2*g, {nf, a, a, adag},{i, i+1, i+2, i+3});
